+++ frontend/src/components/index.ts
@@ -16,6 +16,7 @@ export * from 'components/@shared/ScrollToTop';
 export * from 'components/@shared/ShareButton';
 export * from 'components/@shared/Tooltip';
 export * from 'components/@shared/ToggleButton';
+export * from 'components/@shared/UnderLineText';
 export * from 'components/Logo';
 export * from 'components/ValidationMessage';
 export * from 'components/Header';
@@ -25,7 +26,6 @@ export * from 'components/ChallengeList';
 export * from 'components/CertItem';
 export * from 'components/CheckCircles';
 export * from 'components/Timer';
-export * from 'components/UnderLineText';
 export * from 'components/ChallengeIcon';
 export * from 'components/Profile';
 export * from 'components/EmptyContent';
